fix formatting, share balloon code
